+++ functions/cTab/functions/ATAK/Group/fn_ATAK_onGroupClicked.sqf
