+++ functions/cTab/functions/ATAK/Message/fn_ATAK_msg_Line_Create.sqf
@@ -1,7 +1,7 @@
 params ["_list","_type","_txt"];
 
 private _ctrlMsg = _display ctrlCreate [
-  [configFile >> "RscTitles" >>"ATAK_Message", configFile >> "ATAK_Message"] select _isDialog,
+  [configFile >> "RscTitles" >>"ATAK_Message_Line", configFile >> "ATAK_Message_Line"] select _isDialog,
   -1,
   _list
 ];
